Partie 1 + 2 a corriger avec 42FileChecker

+++ ft_strnew.c
@@ -6,18 +6,17 @@
 /*   By: llacaze <llacaze@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/19 22:54:31 by Lulu              #+#    #+#             */
-/*   Updated: 2017/10/09 19:41:53 by llacaze          ###   ########.fr       */
+/*   Updated: 2017/10/10 01:25:25 by llacaze          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include <stdio.h>
 
-
 char    *ft_strnew(size_t size)
 {
     char    *str;
-    int     i;
+    size_t  i;
 
     i = 0;
     if (!(str = (char *)malloc(sizeof(size))))
